+++ R/geom_half_point.R
@@ -3,7 +3,6 @@
 #' @inheritParams ggplot2::geom_point
 #' @param side The side on which to draw the half violin plot. "l" for left, "r" for right, defaults to "l".
 #' @param transformation An evaluated `position_*()` function yielding a `Position` object with specified parameters to calculate the transformation of the points. Defaults to `ggplot2::position_jitter()`.
-#' @param transformation_params Deprecated. A `list` containing named parameter values for the `transformation` object. Defaults to `list(width = NULL, height = NULL)`. For `ggplot2::PositionJitter`, keyword arguments can be `width`, `height` and `seed`.
 #' @param range_scale If no `width` argument is specified in `transformation_params`, `range_scale` is used to determine the width of the jitter. Defaults to `0.75`, which is half of the allotted space for the jitter-points, whereas `1` would use all of the allotted space.
 #' @importFrom ggplot2 layer position_jitter .pt .stroke
 #' @importFrom grid pointsGrob gpar
@@ -20,7 +19,7 @@ geom_half_point <- function(
   ...,
   side = "r",
   transformation = position_jitter(),
-  transformation_params = list(width = NULL, height = NULL),
+  # transformation_params = list(width = NULL, height = NULL),
   range_scale = .75,
   na.rm = FALSE,
   show.legend = NA,
@@ -36,7 +35,7 @@ geom_half_point <- function(
       params = list(
         side = side,
         transformation = transformation,
-        transformation_params = transformation_params,
+        # transformation_params = transformation_params,
         range_scale = range_scale,
         na.rm = na.rm,
         ...
@@ -74,12 +73,12 @@ GeomHalfPoint <- ggproto(
   draw_group = function(
     data, panel_params, coord, na.rm = FALSE, side = "r", 
     transformation = position_jitter(),
-    transformation_params = list(width = NULL, height = NULL),
+    # transformation_params = list(width = NULL, height = NULL),
     range_scale = .75) {
 
-    if (isFALSE(isTRUE(all.equal(transformation_params, list(width = NULL, height = NULL))))) {
-      warning("Argument deprecated.\n Use `transformation = position_*(params)` instead of passing the params via `transformation_params`")
-    }
+    # if (isFALSE(isTRUE(all.equal(transformation_params, list(width = NULL, height = NULL))))) {
+    #   warning("Argument deprecated.\n Use `transformation = position_*(params)` instead of passing the params via `transformation_params`")
+    # }
     
     if (is.character(data$shape)) {
       data$shape <- translate_shape_string(data$shape)
@@ -88,6 +87,7 @@ GeomHalfPoint <- ggproto(
     xrange <- data$xmax - data$xmin
     x_add  <- (xrange / 4) * switch((side == "r") + 1, -1, 1)
     data$x <- data$x + x_add
+
     # Add Position Transformation
     transformation_params_new <- transformation$setup_params(data)
     transformation_df <- data.frame(
@@ -97,13 +97,14 @@ GeomHalfPoint <- ggproto(
       group = -1
     )
     
+    
     null_width <- FALSE
     name_width <- "width" %in% names(transformation)
     if (isTRUE(name_width)) {
       null_width <- is.null(transformation[["width"]])
     }
     if ((!name_width || null_width) && "width" %in% names(transformation_params_new)) {
-      transformation_params_new$width <- xrange / (2 / range_scale)
+      transformation_params_new$width <- xrange / (4 / range_scale)
     }
     
     null_height <- FALSE
@@ -116,19 +117,19 @@ GeomHalfPoint <- ggproto(
     }
     
     # Add deprecated transformation_params args to new transformation_params list
-    par_idx <- !sapply(transformation_params, is.null)
-    transformation_params_new[names(transformation_params)[par_idx]] <- transformation_params[par_idx]
-    transformation_params <- transformation_params_new
-    
-    # deprecated
-    if ("width" %in% names(transformation_params)) {
-      transformation_params$width <- transformation_params$width %||% xrange / (4 / range_scale)
-    }
-    # deprecated
-    if ("height" %in% names(transformation_params)) {
-      transformation_params$height <- transformation_params$height %||% 
-        ggplot2::resolution(data$point_y[[1]], zero = FALSE) * 0.4
-    }
+    # par_idx <- !sapply(transformation_params, is.null)
+    # transformation_params_new[names(transformation_params)[par_idx]] <- transformation_params[par_idx]
+    # transformation_params <- transformation_params_new
+    # 
+    # # deprecated
+    # if ("width" %in% names(transformation_params)) {
+    #   transformation_params$width <- transformation_params$width %||% xrange / (4 / range_scale)
+    # }
+    # # deprecated
+    # if ("height" %in% names(transformation_params)) {
+    #   transformation_params$height <- transformation_params$height %||% 
+    #     ggplot2::resolution(data$point_y[[1]], zero = FALSE) * 0.4
+    # }
 
     # if (is(transformation, "PositionJitter")) {
     #   transformation_params$width  <- transformation_params$width %||% xrange / (4 / range_scale)
@@ -139,12 +140,12 @@ GeomHalfPoint <- ggproto(
     if (is(transformation, "PositionIdentity") || is(transformation, "PositionJitter")) {
       trans_positions <- transformation$compute_layer(
         transformation_df,
-        transformation_params
+        transformation_params_new
       )
     } else {
       trans_positions <- transformation$compute_panel(
         transformation_df,
-        transformation_params
+        transformation_params_new
       )
     }